+++ SourceCode/Yuconz/src/AnnualReview.java
@@ -1,0 +1,5 @@
+public class AnnualReview extends Document{
+    public AnnualReview(String creator) {
+        super(creator);
+    }
+}