+++ SourceCode/Yuconz/src/Document.java
@@ -1,0 +1,7 @@
+public class Document {
+    private String staffID;
+
+    public Document(String creator) {
+        staffID = creator;
+    }
+}