+++ SourceCode/Yuconz/src/Main.java
@@ -11,10 +11,11 @@ public class Main {
      */
     public static void main(String[] args) {
         AuthServer authServer = new AuthServer();
+        HRDatabase hrDatabase = new HRDatabase();
         authServer.addDetails("hremployee", "kieran", "password");
         authServer.addDetails("employee", "conor", "password");
         authServer.addDetails("employee", "alice", "password");
-        AppController appController = new AppController(authServer);
+        AppController appController = new AppController(authServer, hrDatabase);
         appController.runController();
     }
 }